refactor buncha stuff, abstract storage layer from index, support paginated storage layer, modify interface definitions, more dep injection in main

+++ pb/decode.go
@@ -20,7 +20,7 @@ type OffsetReader interface {
 // Decode reads from r starting at offset. It first gets the little endian encoded length
 // of the Entry data, and then reads the log entry's bytes. Once read it unmarshals them
 // into a log entry, which it returns along with the number of bytes read [4 + len(log-entry-bytes)] unless an error occurs.
-func Decode(r OffsetReader, offset int64) (*Entry, int, error) {
+func Decode(r OffsetReader, offset int64) (*Entry, int64, error) {
 	// read the bytes storing the size of the data
 	sb := make([]byte, 4) // stored as uint32 (4 bytes)
 	_, err := r.ReadAt(sb, offset)
@@ -45,5 +45,5 @@ func Decode(r OffsetReader, offset int64) (*Entry, int, error) {
 		return nil, 0, fmt.Errorf("error: could not unmarshal bytes: %v", err)
 	}
 
-	return &entry, 4 + len(data), nil
+	return &entry, int64(4 + len(data)), nil
 }
